+++ crates/playwright/tests/integration/main.rs
@@ -29,6 +29,7 @@ mod locator;
 mod navigation;
 mod network_route;
 mod page;
+mod page_assertions;
 mod page_event_network;
 mod page_new_methods;
 mod page_properties;